fixing GameMusic

+++ MojFramework/App.h
@@ -47,6 +47,6 @@ private:
 	Sound jazDamaged;
 	Sound startGame;
 	Sound gameMusic;
-	int frameCount;
+	float frameCount;
 	/********************************/
 };
